Extracted RCP product into bootstrap bundle which checks Java version
Problem: if the bundle that defines the RCP product requires a higher
Java version than the version of the current runtime, then the bundle is
not resolved and the application does not start and shows no message to
the user at all.

Solution: extract the product definition into a bootstrap bundle that
can always run with Java 6, check for the version and, if necessary,
show an appropriate message to the end user.

Issue: #196