Fixed lookup of quotes for dates before the available historical quotes
If a quote is required (for example as part of a performance calculation)
for a date before the available historical quotes, the lookup used to
return 0. This, however, leads to zero valuations of position and in turn
to crazy values in the performance calculation. Better to use any quote,
if available.