MaxDD duration: taken longest period even if no new peak has been reached
This change checks if the drawdown duration at the end of the reporting
interval is longer than the max drawdown duration calculated so far. If
so, then this duration is used because it is the longest duration even
if we do not yet know how much longer it will get.

Issue: #240