Use Nebula's CDateTime widget on Linux
On Linux, DateTime does not work very well because of the missing
support for localization. On Mac OS X, CDateTime runs sluggish and does
not render the drop down box correctly. And on Windows, I prefer the
DateTime widget. Because therefore the DateTimePicker will choose
between CDateTime and DateTime depending on the platform.

Issue: #301