Enhanced transactions with units to hold additional values in foreign currency
While this initial commit does not complete the whole story, it adds
the concept of units to the base transaction. A unit is a separate piece
such as fees, taxes, or lump sum values that can be (optionally) in
a foreign currency. It replaces the ForexData object which only kept
the lump sum as foreign currency object.

Issue: #324