Adapted Comdirect Extractor to parse currencies
Currently, the Comdirect extract expects all transactions to be in EUR.
Those are hard-coded delimiters used for parsing.

It would be better to refactor the parsing and use the pattern-matching
approach of the AbstractPDFExtractor. However, that requires the text
with PDFTextStripper#setSortByPosition set to true which in turns
requires to convert the original PDF documents again. Alas, I don't have
the original files.

Issue: #14