Set clearPersistedState based on model.version instead of ini file
If the model changes, the persisted state must be cleared. With this
change the model version is persisted in the preferences, compared upon
startup and, if different, the persisted state is cleared.

Previously, the online update changed the ini file and added the flag.
However, the global ini file must not be changed (possibly no
permissions, signed).