Fix calculation of purchase value when security transfer exist
The purchase value displayed in the 'statement of assets' view is based
on the security position. The calculation within the security position
was not handling security transfers correctly. Why? Because the
positions of the joint portfolio were built out of the positions of the
individual portfolios - from which positions with zero shares had been
removed.

Issue: #371