Implementation of kickbacks WORK-IN-PROGRESS
Kickbacks are implemented as a new transaction type. The GUI and all
pages are loading and working. Currently, kickbacks have the same
properties as dividends, including kickbacks/share. They are also shown
in the category "earning/revenus/Erträge".

An alternative could be to remove the field showing kickbacks/share and
also to show kickbacks in their own category or summing them up in the
category fees (as negative positions).

WORK-IN-PROGRESS: Missing are all the implementations of calculations
within the following folder. There, one needs to discuss first into
which calculations the kickbacks are supposed to go into:
/name.abuchen.portfolio/src/name/abuchen/portfolio/snapshot/security