Remove cancel button from CSV column configuration dialog
Because the original model is edited, the cancel button suggests that
the user can abort the operation. For now, we remove the cancel button.
One could change the implementation in such a way that changes are applied
only when pressing OK, however, I am not sure that it is worth the time.