Add new performance computation
This computation only takes into account the non-neutral transactions.
It also does not compute the total money at the end of the period but
the total sum of all non-neutral transactions.

Pleas note that this commit also changes the computation of the "reduced"
view by using a new common summing method.