Fixed action to convert delivery transactions to buy/sell
When creating a buy or sell transaction out of a delivery transaction,
the currency must match to the currency of the account. With this change,
an error message is thrown before the model is edited (as opposed to
deleting the transaction and failing on re-adding as buy/sell).

We do not attempt to guess an exchange rate and/or combine forex fees
and taxes to fix the transactions. One could allow the user to choose
the target account with a popup. However, this is probably an edge case.