Eurostat HICP as QuoteFeed
According to #1108

This new QuoteFeed is collection the European wide compareable inflation rate (HICP) as JSON stream from Eurostat.

The according reagions are selectable as DropDown.

Known bugs belonging to the region name sorting (eg. Österreich after Zypern) and that the run job @ AbstractQuoteProviderPage is only triggered if an input at security related field was made before. But as this feed does not require initially a symbol the DropDown is not prefilled.

Fix