Degiro: check if a delta exists between net and gross dividend and fix it
Problem: Degiro removes taxes also in foreign currency. If there are small
amounts, rounding differences can create a delta between the gross value
(without taxes) and the net value (with taxes) in transaction currency.
Usually this is 1 Cent.

This change detects if such a delta exist and "fixes" the tax unit with
the delta amount. The disadvantage is that the forex values of the tax
unit does not necessary reflect what was found in the PDF document
anymore.

Issue: #1198