Added 'Holding Period' to the list of available intervals in security chart
Allow the user to display only the interval during which the security
was held. The chart ends today if there are current holdings. If there
is no holding period for a given security, then the last 2 years are
displayed.

To enable holding periods, there are two major code changes: First, the
chart interval is calculated dynamically based on the transactions of
the security. Second, the chart can end earlier than today's date.