Add support for VIAC Tax Refund PDF import
VIAC dividends of type "Ruckerstattung Quellensteuer" seem to only ever
be in CHF. However the underlying securities of the tax refund might be
in different currencies.

In this case the import will currently fail due to the missing exchange rate,
which is not provided in the PDF.

Handle this case by not setting the Security in the transaction, but
only noting the security name in the note of the transaction.

Related #1624 #1514