Do not pack columns in CSV import dialog on GTK / Linux
It seems like the pack() function changes the width of the first element,
as soon as a column is "packed" with a non-default width. This causes
the first column to be set to the width of the table. The reason is
unclear. For now, do not pack the table at all.

Issue: #1744
Co-authored-by: Colinas Maoling <colinas.maoling@t-online.de>
Co-authored-by: Andreas Buchen <andreas.buchen@gmail.com>