Fix error in trail if whole position is transferred (#1951)
Even if the entire position in one portfolio is removed,
we should create a "fractional" entry in the trail
in order to determine a correct initial valuation.

This commit fixes the error seen in #1951; I'm not entirely
certain whether it actually covers all edge cases.

(Additionally, minor spelling fix in trail calculation.)