Perform short pre-optimisation for IRR
In some cases where the NPV curve is very degenerate (#1904),
Newton fails to converge to an IRR. This commit attempts
to remedy the problem by inserting an additional
pre-optimisation phase, where an initial guess in the interval
(-1,0) is derived via interval halving. If no such value is
found, we just guess an IRR of 1.05 for now (maybe this
needs to be improved to cover more cases).
A more accurate IRR value is then obtained by applying the
usual Newton iterations.