(JSON) Quote Feed UI Bugfixes.
AbstractQuoteProvidePage.java line 392 has to be removed, because model.setAttributes(security) already sets all the feed properties (EditSecurityModel.java, line 468), so line 392 adds all the old properties again. This makes no difference, unless the user deletes the content of a text field completely (which sets the property value to null). Due to the filtering in EditSecurityModel.java, line 468 the property gets removed. The duplicated entries now cause a fall back to the old properties instead of using null as value. In the end, the quote feed shown in the preview does not match what is currently set in the GUI.