Show why a quote is unexpected on a certain day
In the data quality pane, a list of dates is shown for which a quote is
present even though the stock exchange was supposed to be closed on that
day, according to the trade calendar in use. As an aid for further
investigation, show which holiday was observed on the relevant day.

E.g., the more helpful text "2021-04-02 (Good Friday)" will be displayed
instead of just "2021-04-02".