Add further NYSE one-time closings
The NYSE was closed for the funeral of former U.S. president Nixon in 1994
and in conjunction with the 9/11 terrorist attacks in 2001. Add these
ad-hoc closings to the NYSE trade calendar.

With these additions, all closings since 1990 are included, according to
the list at https://www.bcm-news.de/wp-content/uploads/closings-nyse.pdf
(which also lists a lot of shorter breaks during trading days).