Fix NYSE New Year moving rule
While as a general rule at NYSE, holidays that happen to fall on a Saturday
are observed on the preceding Friday instead, there is an exception if that
Friday would be the last trading day of a month; which is always the case
for the day before New Year. This exception will next apply for New Year
2022, which will be a Saturday; the holiday will not be moved, and the
stock exchange will be open both on Friday and on Monday. See also
https://www.nyse.com/markets/hours-calendars

Correct the trade calendar for NYSE by removing the moveIf rule for
bringing forward New Year from a Saturday. (The corresponding rule to
move a Sunday New Year to the following Monday is correct.)