Rename “Income & Expenses” to “Payments”
The name “Income & Expenses” (formerly “Earnings”) no longer accurately
describes its functionality, since the results of closed trades are also
available under it, and savings (performance-neutral transactions) soon
will be, too. The shared core of the functions is that payments of various
categories are grouped and summed up, as opposed to changes in value which
are not accompanied by flows of money. Therefore, rename it to “Payments”.

Some discussion about the appropriate terms happened on the forums at
https://forum.portfolio-performance.info/t/9531/9