Do not write stack trace into logs for regular HTTP status codes
Regular users cannot differentiate PP errors (where we need stack trace)
from http error codes (which are typically due to the fact that the URL
is simply wrong, not valid anymore, etc.). We now only write the message
but no stack trace. Plus we add the english text for the error code.