Retain original portfolio in trade transactions
When collecting trades, the transactions making up a trade are stored for
later display. It is sometimes necessary to split such a transaction, e.g.
when a buy transaction belongs to two trades because the bought shares were
later sold in separate transactions. Make sure that the split transaction
still refers to the portfolio in which it was originally done, instead of
any portfolios that it may have been transferred to later.

Issue: https://forum.portfolio-performance.info/t/17792