Delete PDF Author functions in all pdf extractors
Sometimes a 'BankIdentifier' is recognized as PDF author from the PDF extractors.
This can cause the wrong PDF extractor to be selected and errors to occur.
Since the PDF author is not used when extracting PDF's, we do not set it.
The same is true for the help function when PDF debug is created.