Optimization of markers and font in security chart
The dividend texts can be stacked up to three times.
As soon as an overlapping of the texts is possible, no text is written or is suspended until there is free space for the text again.

The position of the text of the highest and lowest price has been swapped.

The respective offset value of the markers was removed and adapted by the respective font size and/or by the marker size.