Rebuild OnVista Bank PDF-Importer
Rebuild OnVista PDF-Importer
Merge buy and sell transaction
Merge delivery_inbound and delivery_outbound transaction
Fix missing time in dateTime
Fix dateTime with year of two and four digits
Fix incorrect extraction of the name from the security
Fix calculation of withholding taxes
Fix of incorrect regEx for shares and bonds
Fix some other wrong transactions
Deleting duplicate testcases and testfiles
https://forum.portfolio-performance.info/t/pdf-import-von-onvista-bank/2076/83
Integration in the PDFExtractorUtils.java (fee's and tax's)

Deletion of the "deposit statement" function.
This leads to deposits which take place through other PDF's and deposit them on a wrong date.
This PDF is not suitable for the use of PP and leads to incorrect postings in total.