Rebuild OnVista PDF-Importer
Merge buy and sell transaction
Merge delivery_inbound and delivery_outbound transaction
Fix missing time in dateTime
Fix dateTime with year of two and four digits
Fix incorrect extraction of the name from the security
Fix calculation of withholding taxes
Fix of incorrect regEx for shares and bonds
Fix some other wrong transactions
Deleting duplicate testcases and testfiles
Delete addDepositTransaction()
https://forum.portfolio-performance.info/t/pdf-import-von-onvista-bank/2076/83
Integration in the PDFExtractorUtils.java (fee's and tax's)

Information:
Deletion of the "deposit statement" (Depotauszug) function.
This results in the recorded transactions all being posted to the date the deposit statement was created. e.g. (deposit statement as at 31.12.2010). So if PDF's are imported where individual transactions such as purchase, sale (...) are recorded, double bookings will occur.
This PDF is not suitable for the use of PP and leads to false entries in total.