Fixes wrong part being activated when switching parts using the tab folder
For whatever reason, when the CTabFolder (used for the part stack)
tries to "fix" the focus, the chart legend forces a selection event if
he first tab is currently not selected. The selection event overrides
the selection event from picking the tab. This results in "jumpy"
behavior when selecting tabs (the user clicks on the third one, but
the first is activated)

fix: never force focus on the legend but focus the plot area instead

Issue: https://forum.portfolio-performance.info/t/auswahl-verschiedener-dateien-springt-beim-wechsel-zur-ersten-datei/18598
Issue: https://www.wertpapier-forum.de/topic/38306-portfolio-performance-mein-neues-programm/page/168/?tab=comments#comment-1472792