Fixed default sort order for new views + restore saved sorting behavior
This changes updates the default sort order for the existing views:
for name columns it is ascending, for date column it is descending.

In order to restore tables exactly the way the user has previously
sorted the table, the table configuration in ShowHideColumnHelper
introduces two new values for ascending and descending to differentiate
them from the old ones. After some time has passed by, there should be
no local table configuration stored in the workspace of the user and we
can remove the legacy values.

Issue: #2580