Refactored #processFeeEntries and added #asBigDecimal method
By moving the instanceof checks into the PDFExtractorUtils, it is
possible to reuse the PDFExtractorUtils directly (for example if one
already has values and does not want to work with strings).

Added a asBigDecimal method to have more readable code. Previously,
we used the asExchangeRate method but the naming implies something
different

Issue: #2682