Print tax or fees in case no meaningful portfolio rate can be calculated
The tax or fee rate calculates the fees or taxes in relation to the
gains in a given period. However, if the portfolio has a negative
performance or if there are only tax or fee refunds, then the rate
does not make sense. This change prints instead the tax or fee amount.

Issue: https://www.wertpapier-forum.de/topic/38306-portfolio-performance-mein-neues-programm/?do=findComment&comment=1504176