Fix 2000-10-03 in trade calendar for Germany
In 1999, Deutsche Börse announced that it would trade on certain public
holidays, including German Unity Day (October 3), from 2000. It turns out,
however, that due to resistance from trade unions, this was not actually
implemented for German Unity Day that year; the exchanges were in fact
closed on 2000-10-03. It was only from 2001 that Germany Unity Day
became a trading day. Change the trade calendar for Germany accordingly.

https://www.spiegel.de/wirtschaft/boerse-dax-schliesst-mit-gewinn-a-96375.html
https://www.verdi.de/presse/pressemitteilungen/++co++ee8206a4-4708-11d9-530a-003048429d94