Fix 2007-05-28 in trade calendar for Germany
Between 2000 and 2014, Whit Monday (a public holiday) was generally a
trading day at Deutsche Börse’s exchanges, with the motivation that foreign
exchanges (in particular NYSE and LSE) continued trading, too. However, in
2007, Whit Monday fell on 2007-05-28, the last Monday in May; it therefore
coincided with Memorial Day in the U.S., a closing day for NYSE, and with
the Spring Bank Holiday in the UK, which made that point moot. Therefore,
the German exchanges remained closed on that day as well. Change the trade
calendar for Germany accordingly.

https://www.xetra.com/resource/blob/311588/db9b96a1cc5ab63a74657ec54072e796/data/handelskalender-2007.pdf
https://www.handelsblatt.com/finanzen/maerkte/boerse-inside/verzicht-auf-feiertagshandel-boerse-will-naechstes-jahr-an-pfingsten-nicht-oeffnen/2682906.html
https://www.faz.net/aktuell/finanzen/duenn-besetztes-boersenparkett-feiertagshandel-ohne-elan-1438131.html