Distinguish between Christmas, Boxing Day, etc.
Many countries celebrate holidays on December 25 and/or 26, but there are
different traditions, and the days may be called 1st and 2nd day of
Christmas, or Christmas and Boxing Day, or Christmas and St. Stephen's Day,
and be celebrated in different ways. More clearly distinguish between them
in strings and trade calendars.