Modify Baader Bank PDF-Importer to support new transactions (#2849)
Remove buy, sell and dividend transactions in depot account statement

In the periodic account statement posting, purchase, sale and dividends
were previously also read out. This works, but is not correct.
No fees or taxes are recorded for these entries. We therefore limit ourselves
only to deposits and deliveries, as well as any fees for the account.
The purchases, sales and dividends are recorded in other documents.

Issue: #2849 #2849
Issue: https://forum.portfolio-performance.info/t/pdf-import-von-baader-bank-scalable-capital/2057/74