Make the weekend in trade calendars configurable
Trade calendars up to now have a hard-wired notion of weekend (always
comprising Saturday and Sunday), except for the empty calendar, which is
equally hard-wired to not have a weekend at all, and the first-of-month
calendar, for which the concept of a weekend doesn’t apply. This makes the
code harder to maintain than necessary and poses a problem if further trade
calendars with nonstandard weekends are to be added at some future point
(e.g., for the stock exchange of Tel Aviv, where the weekend is Friday and
Saturday).

Make the weekend a configurable property of the trade calendar and remove
the specialcasing for the empty calendar.