Extract Security exchange from DeGiro transaction PDFs and Display it
DeGiro includes the 3 character exchange of the security for BuySell
transactions in the transaction PDFs.

This commit adds exchange to the Security model class and extracts the
exchange by modifying the regexes of the existing 15 DeGiro extraction
sections of DegiroPDFExtractor.

The exchange is displayed in a new column in the import dialog
(ReviewExtractedItemsPage) and in a new column in the Securities view
(SecuritiesTable).

Asserts for the exchange have been added to all JUnit tests for DeGiro
Transaction PDFs in DegiroPDFExtractorTest.