Fix observance of Swiss national holidays
Between 2001 and 2005, the Swiss stock exchange (then SWX, now SIX) did not
observe the Swiss national holiday (1 August), continuing trading on that
day instead. (For 2001, this was decided on rather short notice, as the day
was still marked a non-trading day on the website as of April 2001; no
longer in June 2001, however.) Fix this in the SIX trade calendar.

https://web.archive.org/web/20000301013300/http://www.swx.com:80/market/tradcal2000.pdf
https://web.archive.org/web/20010423015445/http://www.swx.com:80/market/trading_calendar_en.html
https://web.archive.org/web/20010619200842/http://www.swx.com:80/market/trading_calendar_en.html
https://web.archive.org/web/20030411001712/http://www.swx.com/market/tradcal2002.pdf
https://web.archive.org/web/20040731224140/http://www.swx.com/market/tradcal2003.pdf
https://web.archive.org/web/20040723162143/http://www.swx.com/market/tradcal2004.pdf
https://web.archive.org/web/20060324043848/http://www.swx.com/trading/information/calendar/2005/grid_en.html
https://web.archive.org/web/20060324041803/http://www.swx.com/trading/information/calendar/2006/grid_en.html