Fix SIX/SWX calendar for 2002
In 2002, the Swiss stock exchange (then SWX, now SIX) followed the trend of
other stock exchanges at the time and stopped observing a number of less
universal holidays, namely, Berchtoldstag, Ascension and Whit Monday.
(There were also plans for 2001, 2003 and 2004 to abolish Christmas Eve
and/or New Year’s Eve, and while these made it to some versions of the
trading calendar, they were not followed through in the end.) In 2003, the
exchange abandoned the attempt and returned to the usual holidays, with the
exception of the national day (see the previous commit).

https://web.archive.org/web/20030411001712/http://www.swx.com/market/tradcal2002.pdf
https://web.archive.org/web/20030411061250/http://www.swx.com/market/tradcal2003.pdf
https://web.archive.org/web/20040731224140/http://www.swx.com/market/tradcal2003.pdf
https://web.archive.org/web/20040723162143/http://www.swx.com/market/tradcal2004.pdf