Add FinnHub.io as Security Search provider when adding instruments
Added FinnhubSearchProvider and FinnhubSymbolSearch to use the free
Finnhub.io Symbol Lookup API to search for a security by symbol or ISIN
when adding a new investment vehicle. This works in the same way as the
existing YahooSearchProvider.

If the user has not configured a Finnhub.io API key then the search API
is not called. This avoids 401 Unauthorized for those users.

https://finnhub.io/docs/api/symbol-search

Issue: #2878
Signed-off-by: Egal <31034254+ah4c@users.noreply.github.com>
[squashed commits; rebased to master]
Signed-off-by: Andreas Buchen <andreas.buchen@gmail.com>