Add PP Version to PDF Handler
Sometimes the question of which version of PP is installed at the client arises with a PDF debug. With this change I just want to capture which version of PP is installed. It happens that the end user is several versions behind because something went wrong with the PP update routine.