Added EODHistoricalData.com as Search Provider
Added EODHistoricalDataSearchProvider to use the free
EODHistoricalData.com Symbols Lookup API to search for a security by
ISIN, ticker symbol or name when adding a new investment vehicle. This
works in the same way as the existing Search Providers and provides
additional results.

If the user has not configured a EODHistoricalData.com API key then the
search API is not called. This avoids authentication errors for those
users.

The EODHistoricalData.com Symbols API returns:

* Code - the ticker symbol
* Exchange - EODHistoricalData-specific code
* Name
* Type
* Country
* Currency - 3 characters
* ISIN
* previousClose
* previousCloseDate

For example:

[
{
"Code": "BBVAI",
"Exchange": "MC",
"Name": "Accion IBEX 35 Cotizado Armonizado FI",
"Type": "ETF",
"Country": "Spain",
"Currency": "EUR",
"ISIN": "ES0105336038",
"previousClose": 8.264,
"previousCloseDate": "2022-07-08"
}
]

The Code does not include an exchange suffix. Fortunately, the Exchange
code can be appended to the Code to get a ticker symbol with exchange.
Note that Exchange is the EODHistoricalData.com code, which is often the
same as the Yahoo code but not always. refer to
https://eodhistoricaldata.com/financial-apis/list-supported-exchanges/
for the mapping between EODHistoricalData.com exchange code and the
4-character MIC.

The EODHistoricalData.com API key is provided to
EODHistoricalDataSearchProvider by Preference2EnvAddon in the same
setEODHistoricalDataApiKey() method where it is provided to the existing
EODHistoricalDataQuoteFeed.

Issue: #2905
Signed-off-by: Egal <31034254+ah4c@users.noreply.github.com>
[reformatted to PP code style; rebased to master]
Signed-off-by: Andreas Buchen <andreas.buchen@gmail.com>