Add year selection to the calendar preview
On the calendar preference page, a preview of the holidays in the chosen
calendar was already shown. It was, however, restricted to the current
year. Since it may be relevant to see the holidays of the next year in
particular, but sometimes even of other years, add a control through which
the user may specify the year for the calendar preview.