Updated rendering of slices in pie/donut charts that fall below 1 degree
The GC#fillArc function only allows angle and arc as int. Therefore
slices that are below 1 degree are difficult to draw. Previously, those
slices were rounded up to 1 degree. However, that resulted in an
"overrun" of the circle. This commit inserts gaps for each consecutive
series of slices below 1 degree to indicate the additional slices to the
user.