Create a .gitattributes file
To ensure that no text files with CR+LF line endings make it into the
repository, add a .gitattributes file which tells Git that text files are to
be autoconverted to LF-only when on Windows. This applies to the PP
repository, whatever the local settings of the user creating the commit are.