Modify Quirin PDF-Importer to support new transaction
https://forum.portfolio-performance.info/t/pdf-import-von-quirin-privatbank-ag/21748/19
https://forum.portfolio-performance.info/t/pdf-import-von-quirin-privatbank-ag/21748/21

In the PDF debug of the portfolio statement, purchases and sales as well as dividends were recorded. This is analogous to the Baader Bank PDF importer not okay. In the portfolio statement, taxes and fees are not correctly reflected or the exchange rates are missing. These were hereby removed, as there are individual statements with the correct data.

Only deposits and deliveries are handled via portfolio statements.