Additional limit of securities in BookmarkMenu
As wrote here: a56614c8a5fafd3c7c552f9c51edd861d86a3132

If you want limit then we have to add the ``.stream().limit(10)`` here, too. Or Alternatively limit the securities at central place in constructor.

But I think, implementing a constant and (for the user) hidden limit can cause confusion and unneccessary error reports (sooner or later).
So I will try to propose a PR soon were there is no constant limit, but a confirmation message, the user have to confirm if the limit is exceeded. So the users still can use the function without limitation but with extra click.