Added toMutableList collector to mark the need for a mutable list
It's useful in cases where we need to modify the resulting list, which
is not possible with the immutable list returned by the Stream#toList
method. We use this method to differentiate between cases where a
mutable list is necessary and to prevent the IDE from suggesting the use
of Stream#toList.

Issue: #3197