[DRAFT] Refactored non-importable items to use failure message instead
This commit replaces the use NonImportableItem with a regular item that
has a failure message. This way all item types can be marked as failure
while presenting each type specific data to the user.

Extractors can use this feature by storing a FAILURE message in the
transaction context and setting the message when wrapping the item.

Issue: #3119