Fix ClientClassificationFilter re. interest w/ tax
When, in commit 5795ad2d5028abde855be66e0af4add565822ed7, interest
transactions were enabled to have taxes, the ClientClassificationFilter
wasn’t updated accordingly, and taxes contained in interest transactions
were ignored: Only the net amount of the interest was considered. Ensure
that the filtered client receives the gross amount of interest together with
a withdrawal for the tax amount.