Extend input field calculations to subtractions
It was already possible to do small additions in an input field (e.g., the
input “10.01+23.45” was interpreted as 33.46). Extend this to also allow
subtractions, such as “33.46-23.45”, or indeed any combination such as
“-10.01+23.45+47.11-0.03”. To this end, split the input on plus or minus
(instead of only on plus), using a lookahead to retain any minus signs, and
only check the total result for an undesired negative value, instead of each
part separately.

Issue: https://forum.portfolio-performance.info/t/9749/2