Improvement in creation of the PDF debug
Add dialog title

Adding an information text when creating a PDF-
Troubleshooting

Added a mouse listener which, when you double-click on a string, replaces it with a random string with exceptions (currency, ISIN, date formats, time formats).
It will replace letters with letters or numbers with numbers.

Added the function "containsCurrencyCode" to check if the marked string is a currency for anonymization

Added the three "``" at the beginning and at the end of the PDF debug for preformatting in the PP forum