Do not abort all quote jobs for a missing API key
Upon trying to retrieve quotes from Alpha Vantage, when an API key wasn’t
configured, the AlphavantageQuoteFeed would throw an exception. This caused
the parallel jobs for retrieving quotes to be cancelled, including for
securities that used a different source. Thus, automatic download of quotes
was as good as disabled if at least one security was set to use Alpha
Vantage, but the API key was not set.

Instead of throwing an exception, report the error without affecting other
quote download jobs.

Issue: #3300