Use adaptive widget size in the currency converter
In the currency converter tab, the size of the widget that shows the sources
of the exchange rate was fixed to 420 pixels wide and 200 pixels high. On a
modern, high-resolution screen, this isn’t much, and most of the information
was only visible after scrolling. Instead, use a size that depends on the
current font.