Expand entries in the exchange rate source view
In the currency converter tab, the sources for the exchange rate are presented
in a tree view. The elements of the tree were collapsed by default. This made
it tedious to actually review the information, since one had to manually expand
often multiple elements. Change the behaviour to have all elements expanded by
default.