Improve PWP Leeway to set unknow exchange markets
If the exchange market is not present in the ExchangeLabels list, then
we insert it. If we go through the API and make a query of the unknown
exchange market, then we get a response back in the form of an array,
but we don't get any results via the new "code".

Add leeway.US -> "OTC Market Group"
Fix NPE, when creating a new empty security the ticker symbol is searched.

Issue: #3364
Signed-off-by: Nirus2000 <webmaster@nirus-online.de>
[removed unknown exchange lagel; rebased to master]
Signed-off-by: Andreas Buchen <andreas.buchen@gmail.com>