Fixed newly created security being added to watchlist of different file
The problem was that the posted event lacked information about the
client context in which the security was created. As a result, all open
watchlists were erroneously adding the security to themselves.

This change addresses the issue by creating an event that includes the
client context. Additionally, the change is using the already existing
event code for editing an existing security.

Issue: #3452 #3453
Co-authored-by: OnkelDok <90478568+OnkelDok@users.noreply.github.com>
Co-authored-by: Andreas Buchen <andreas.buchen@gmail.com>