Revert "Added all years to payments/year tooltip"
This reverts commit 5a4d4eddafdc42374e81ed1ed97011808fb09ddf.

Background: When showing all years, the tool tip is clutter with
unnecessary information, for example even if one looks at the current
year only, then instruments without payments in the given year are
listed. If one wants to have all years, then there is already a table
for that purpose.