[PDF] [Deutsche Bank] Fix detection of currency for account statement
The latest account statements cannot be imported because the currency
was not detected. Apparently the latest document (can) introduce a line
break between account number and final balance.

With this change, the currency is retrieved from the header of the
document. However, to enable this we need to check for two lines.