Revision ComDirect PDF-Importer
Closes #3553
Closes #3313
Closes #1131

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/229

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/236

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/242

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/254

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/270

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/276

https://forum.portfolio-performance.info/t/pdf-import-von-comdirect/1647/283

ComDirect provides two documents for the transaction.
The security transaction and the taxes treatment.
Both documents are provided as one PDF or as two PDFs.

The security transaction includes the fees, but not the correct taxes
and the taxes treatment includes all taxes (including withholding tax),
but not all fees.

Therefore, we use the documents based on their function and merge both documents,
if possible, as one transaction.

Always import the securities transaction and the taxes treatment for a correct transaction.
Due to rounding differences, the correct gross amount is not always shown in the securities transaction.