Adds the option for non cash-effective dividend payments
This commit allows to mark dividends as non cash-effective. It can be
used, e.g., for spin-offs or stock dividends, where the paid amount is
directly reinvested in a stock and not paid in cash. The marked
dividends can be excluded in the payment view to keep a clean view of
the dividend development. All non cash-effective dividends can be
filtered and listed in the transaction view.