CSVAccountTransactionExtractor: Do not blindly ignore amount sign
CSV importer's existing mental model is that each transaction would be
explicitly categorized (via a "type" column) as Buy or Sell. It then
just ignores the sign of transaction amount (by calling Math.abs()).

This assumption is not realistic to cover all the cases. Another way
is to have a generic "security transaction" type, and the sign of the
amount signifies the actual operation, e.g. negative for Buy, positive
for Sell.

This change allows to handle such CSV statements, by "inverting" the
configured transaction type. For example, if given type column value
is mapped to "BUY", but amount is negative, it is turned into "SELL".