Improve matching of securities during import: prefer active instruments
If the client contains multiple securities with identical identifier,
then the import must used to ignore the security completely. With this
change, the code now looks at the retired flag of the instrument. If
there is a conflict, but only one of the securities is active, then
transactions are imported to the active security.

Issue: #3592
Co-authored-by: Jan <17313391+rcjan@users.noreply.github.com>
Co-authored-by: Andreas Buchen <andreas.buchen@gmail.com>