TradesTableViewer: Actually show cash account in trade breakdown tooltip
A tooltip with per-transaction breakdown of a trade, shown when hovering
over a trade entry, has a column named "Cash Account". However, what's
actually shown there is *portfolio* account. So, either column title is
wrong, or value shown is wrong. Showing portfolio account wouldn't make
much sense, as a trade always belongs to a single portfolio account,
so value shown would be invariantly the same in each row. And user can
see portfolio account in trade entry (if corresponding column is enabled),
so duplicating it in the tooltip doesn't make much sense. Thus, it seems
that the corresponding cash account was indeed intended to be shown, so
let's make it so.